test: introduce tests
This change introduces a simple test.sh script, which uses mocked `ip`
command. It lists two interfaces (lo and eth0) for `ip a` but it just
prints called parameters otherwise. The test script compares if the list
of called ip commands equals to expected output.

Only load and flush functions are tested currently.